requirements.txt
autopep8==1.6.0
discord==1.7.3
discord.py==1.7.3
python-decouple==3.5
requests==2.26.0
tabulate==0.8.9
tekore==4.1.0
yarl==1.7.2
youtube-dl==2021.12.17
PyNaCl==1.5.0
